# swagger.yaml
openapi: 3.0.0
info:
  title: Cinefy API
  description: Account registration, email verification, authentication, account management, and password reset API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      description: Accounts must be verified before authenticating. Access tokens expire after 15 minutes; refresh tokens expire after 7 days.
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "jason@example.com"
                password:
                  type: string
                  example: "pass123"
              required:
                - email
                - password
      responses:
        "200":
          description: The account email, a JWT access token and a refresh token cookie
          headers:
            Set-Cookie:
              description: "`refreshToken` cookie. HttpOnly and SameSite=Lax; Secure is enabled in production."
              schema:
                type: string
                example: refreshToken=<opaque-token>; Path=/; Expires=Tue, 16 Jun 2020 09:14:17 GMT; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: "jason@example.com"
                  jwtToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
        "400":
          description: The email or password is incorrect, or the request body failed validation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email or password is incorrect"
  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      description: The refresh token is sent and returned via cookies. Each refresh token expires after 7 days and is rotated when used.
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refreshToken
          description: Required refresh-token cookie.
          required: true
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
      responses:
        "200":
          description: The account email, a JWT access token and a new refresh token cookie
          headers:
            Set-Cookie:
              description: "`refreshToken` cookie. HttpOnly and SameSite=Lax; Secure is enabled in production."
              schema:
                type: string
                example: refreshToken=<opaque-token>; Path=/; Expires=Tue, 16 Jun 2020 09:14:17 GMT; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: "jason@example.com"
                  jwtToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
        "400":
          description: The refresh token is invalid, revoked or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid token"
  /accounts/logout:
    post:
      summary: Log out the current session
      description: Revokes the refresh token from the cookie when active and clears the cookie. Existing access JWTs remain valid until they expire.
      operationId: logout
      parameters:
        - in: cookie
          name: refreshToken
          description: The refresh token cookie to revoke. Optional; logout is idempotent.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The refresh token cookie was cleared
          headers:
            Set-Cookie:
              description: Clears the `refreshToken` cookie.
              schema:
                type: string
                example: refreshToken=; Path=/; HttpOnly; SameSite=Lax; Max-Age=0
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
  /accounts/me:
    get:
      summary: Get the authenticated account's details
      description: Returns basic details for the account identified by the bearer token.
      operationId: getCurrentAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the authenticated account
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  gender:
                    type: string
                    example: "female"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  phone:
                    type: string
                    example: "+15551234567"
                  department:
                    type: string
                    enum: [Finance, HR, Engineering, Administration, Operation, Marketing]
                    example: "Engineering"
                  status:
                    type: string
                    enum: [active, inactive]
                    example: "active"
                  role:
                    type: string
                    example: "Admin"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    format: date-time
                    example: "2020-05-08T03:11:21.553Z"
                  isVerified:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      description: Admin users can revoke the tokens of any account, regular users can only revoke their own tokens.
      operationId: revokeToken
      security:
        - bearerAuth: []
      parameters:
        - in: cookie
          name: refreshToken
          description: The refresh token cookie is used unless a token is supplied in the request body, in which case the body token takes precedence.
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550"
      responses:
        "200":
          description: The refresh token was successfully revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token revoked"
        "400":
          description: The refresh token is missing or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is required"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /accounts/register:
    post:
      summary: Register a new user account and send a verification email
      description: The first account registered in the system is assigned the `Admin` role, other accounts are assigned the `User` role. Duplicate registrations return the same success response to avoid disclosing whether an email is registered.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gender:
                  type: string
                  description: Required gender value; the API does not constrain this to a fixed set.
                  example: "female"
                firstName:
                  type: string
                  example: "Jason"
                lastName:
                  type: string
                  example: "Watmore"
                email:
                  type: string
                  format: email
                  example: "jason@example.com"
                phone:
                  type: string
                  example: "+15551234567"
                department:
                  type: string
                  enum: [Finance, HR, Engineering, Administration, Operation, Marketing]
                  example: "Engineering"
                password:
                  type: string
                  minLength: 6
                  example: "pass123"
                confirmPassword:
                  type: string
                  description: Must exactly match password.
                  example: "pass123"
                acceptTerms:
                  type: boolean
                  enum: [true]
              required:
                - gender
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - acceptTerms
      responses:
        "200":
          description: Registration was accepted and a verification email was sent, or an already-registered email notification was sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration successful, please check your email for verification instructions"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email after registration
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "3c7f8d9c4cb348ff95a0b74a1452aa24fc9611bb76768bb9eafeeb826ddae2935f1880bc7713318f"
              required:
                - token
      responses:
        "200":
          description: Verification was successful so you can now login to the account
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Verification successful, you can now login"
        "400":
          description: The token is invalid or the request body failed validation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Verification failed"
  /accounts/forgot-password:
    post:
      summary: Submit an email address to request a password reset; reset tokens expire after 24 hours
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "jason@example.com"
              required:
                - email
      responses:
        "200":
          description: The request was received. A reset email is sent only when the email belongs to an account; the response is identical either way.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Please check your email for password reset instructions"
        "400":
          $ref: "#/components/responses/ValidationError"
  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email after submitting to the /accounts/forgot-password route
      operationId: validateResetToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "3c7f8d9c4cb348ff95a0b74a1452aa24fc9611bb76768bb9eafeeb826ddae2935f1880bc7713318f"
              required:
                - token
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is valid"
        "400":
          description: The token is invalid or the request body failed validation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid token"
  /accounts/reset-password:
    post:
      summary: Reset the password for an account
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "3c7f8d9c4cb348ff95a0b74a1452aa24fc9611bb76768bb9eafeeb826ddae2935f1880bc7713318f"
                password:
                  type: string
                  minLength: 6
                  example: "newPass123"
                confirmPassword:
                  type: string
                  description: Must exactly match password.
                  example: "newPass123"
              required:
                - token
                - password
                - confirmPassword
      responses:
        "200":
          description: Password reset was successful so you can now login to the account with the new password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successful, you can now login"
        "400":
          description: The token is invalid or the request body failed validation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid token"
  /accounts:
    get:
      summary: Get a list of all accounts
      description: Restricted to admin users.
      operationId: getAllAccounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    gender:
                      type: string
                      example: "female"
                    firstName:
                      type: string
                      example: "Jason"
                    lastName:
                      type: string
                      example: "Watmore"
                    email:
                      type: string
                      example: "jason@example.com"
                    phone:
                      type: string
                      example: "+15551234567"
                    department:
                      type: string
                      enum: [Finance, HR, Engineering, Administration, Operation, Marketing]
                      example: "Engineering"
                    status:
                      type: string
                      enum: [active, inactive]
                      example: "active"
                    role:
                      type: string
                      example: "Admin"
                    createdAt:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
                    isVerified:
                      type: boolean
                      example: true
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new account
      description: Restricted to admin users.
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gender:
                  type: string
                  example: "female"
                firstName:
                  type: string
                  example: "Jason"
                lastName:
                  type: string
                  example: "Watmore"
                email:
                  type: string
                  format: email
                  example: "jason@example.com"
                phone:
                  type: string
                  example: "+15551234567"
                department:
                  type: string
                  enum: [Finance, HR, Engineering, Administration, Operation, Marketing]
                  example: "Engineering"
                status:
                  type: string
                  enum: [active, inactive]
                  example: "active"
                password:
                  type: string
                  minLength: 6
                  example: "pass123"
                confirmPassword:
                  type: string
                  description: Must exactly match password.
                  example: "pass123"
                role:
                  type: string
                  enum: [Admin, User]
              required:
                - gender
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - role
      responses:
        "200":
          description: Account created successfully, verification is not required for accounts created with this endpoint. The details of the new account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  gender:
                    type: string
                    example: "female"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  phone:
                    type: string
                    example: "+15551234567"
                  department:
                    type: string
                    enum: [Finance, HR, Engineering, Administration, Operation, Marketing]
                    example: "Engineering"
                  status:
                    type: string
                    enum: [active, inactive]
                    example: "active"
                  role:
                    type: string
                    example: "Admin"
                  createdAt:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  isVerified:
                    type: boolean
                    example: true
        "400":
          description: Validation failed or the email is already registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email 'jason@example.com' is already registered"
        "409":
          $ref: "#/components/responses/ConflictError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ConflictError:
      description: An account with this email already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: The request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
...
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Validation error: \"email\" must be a valid email"